requirements.txt
openai
google-generativeai
crewai
langchain
chromadb
streamlit
python-dotenv
requests

